requirements_clean.txt
# Enhanced Trading Bot Dependencies
# Optimized for both local development and Railway deployment

# Core trading dependencies (required)
alpaca-trade-api==3.0.2
pandas==2.0.3
numpy==1.24.4

# HTTP and async (required)
requests==2.31.0

# Database (optional, for logging)
psycopg2-binary==2.9.7

# Web dashboard (optional)
flask==2.3.3
flask-cors==4.0.0
plotly==5.17.0

# Development tools (optional)
python-dotenv==1.0.0
